mistakes solved
