+++ src/App.css
@@ -16,4 +16,5 @@
   --primary-filters-area-color: #ebf0f5; /* --label-color */
   --primary-button-content-color: #ffffff;
   --danger-button-content-color: #ffffff;
+  --completed-text-color: #0fb864;
 }
