+++ src/pages/OrdersPage/OrdersPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './OrdersPage.module.css';
 import { Pageheader } from './modules/Pageheader/Pageheader';
-import { TableContainer } from './modules/Table/TableContainer';
+import { TableContainer } from './modules/TableContainer/TableContainer';
 import { Filters } from './modules/Filters/Filters';
 
 export function OrdersPage() {
