+++ src/pages/OrdersPage/modules/Filters/FilterStatus/FilterStatus.jsx
@@ -9,7 +9,7 @@ import {
 } from '../../../../../shared/ui';
 import styles from './FilterStatus.module.css';
 
-const STATUS_FILTERS = {
+export const STATUS_FILTERS = {
   new: 'Новый',
   calculation: 'Рассчет',
   confirmed: 'Подтвержден',
