--- src/pages/OrdersPage/modules/Table/TableBody/TableBody.module.css
@@ -1,80 +1,0 @@
-._ {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  background-color: var(--primary-background-color);
-  overflow: auto;
-}
-
-.row:hover {
-  background-color: var(--secondary-color);
-}
-
-.row:last-child {
-  border-bottom: none;
-}
-
-.cell {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.75rem 0.5rem;
-  box-sizing: border-box;
-}
-
-.cell_filtering {
-  padding: 0;
-}
-
-.cellCheckbox {
-  padding: 0.75rem 0.5rem;
-}
-
-.cell_canseled {
-  opacity: 0.5;
-}
-
-.text_completed {
-  color: var(--completed-icon-color);
-}
-
-.cell:nth-child(2) {
-  width: 7.2%;
-}
-
-.cell:nth-child(3) {
-  width: 11.7%;
-}
-
-.cell:nth-child(4) {
-  width: 11.7%;
-}
-
-.cell:nth-child(5) {
-  width: 6.4%;
-}
-
-.cell:nth-child(6) {
-  width: 8.5%;
-}
-
-.cell:nth-child(7) {
-  width: 52.4%;
-}
-
-::-webkit-scrollbar {
-  width: 0.25rem;
-}
-
-::-webkit-scrollbar-track {
-  margin: 0.25rem 0;
-  width: 0.25rem;
-  border-radius: 0.5rem;
-  background: rgba(0, 0, 0, 0.1);
-}
-
-::-webkit-scrollbar-thumb {
-  width: 0.25rem;
-  border-radius: 0.5rem;
-  background: rgba(0, 0, 0, 0.34);
-}
