--- src/pages/OrdersPage/modules/Table/TableContainer.jsx
@@ -1,8 +1,0 @@
-import React, { useContext } from 'react';
-import { FilterContext } from '../../../../store/FilterContext';
-import { Table } from './Table';
-
-export function TableContainer() {
-  const { orders } = useContext(FilterContext);
-  return <Table orders={orders} />;
-}
