--- src/pages/OrdersPage/modules/Table/TableFooter/TableFooter.module.css
@@ -1,45 +1,0 @@
-._ {
-  display: flex;
-  justify-content: space-between;
-  padding: 0.5rem 1rem;
-  background-color: var(--primary-filters-area-color);
-}
-
-.buttons {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.bunch {
-  color: var(--primary-color);
-}
-
-.overlayButton:hover {
-  background-color: var(--primary-color);
-}
-
-.overlayButton:hover span {
-  color: var(--primary-background-color);
-}
-
-.dropdown {
-  position: relative;
-}
-
-.dropdown_delete {
-  position: absolute;
-  left: 0;
-  top: -7.25rem;
-  display: flex;
-}
-
-.pages {
-  display: flex;
-  gap: 1.5rem;
-}
-
-.footer__pagination {
-  display: flex;
-  gap: 0.25rem;
-}
