--- src/pages/OrdersPage/modules/Table/TableHeader/TableHeader.module.css
@@ -1,50 +1,0 @@
-._ {
-  background-color: var(--primary-color);
-}
-
-.wrapper {
-  display: flex;
-  padding: 0 0.5rem;
-}
-
-.cell_filtering {
-  padding: 0;
-}
-
-.cellCheckbox {
-  padding: 0.75rem 0.5rem;
-}
-
-.text {
-  color: var(--primary-filters-area-color);
-}
-
-.icon {
-  width: 1rem;
-  height: 1rem;
-  fill: var(--primary-filters-area-color);
-}
-
-.cell:nth-child(2) {
-  width: 7.2%;
-}
-
-.cell:nth-child(3) {
-  width: 11.7%;
-}
-
-.cell:nth-child(4) {
-  width: 11.7%;
-}
-
-.cell:nth-child(5) {
-  width: 6.4%;
-}
-
-.cell:nth-child(6) {
-  width: 8.5%;
-}
-
-.cell:nth-child(7) {
-  width: 52.4%;
-}
