+++ src/pages/OrdersPage/modules/TableContainer/TableContainer.module.css
@@ -1,0 +1,130 @@
+.wrapper {
+  display: flex;
+  padding: 0 0.5rem;
+}
+
+.cell {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.75rem 0.5rem;
+  box-sizing: border-box;
+}
+
+.cell_filtering {
+  padding: 0;
+}
+
+.cell_sorting {
+  padding: 0;
+}
+
+.cellCheckbox {
+  padding: 0.75rem 0.5rem;
+}
+
+.text {
+  color: var(--primary-filters-area-color);
+}
+
+.icon {
+  width: 1rem;
+  height: 1rem;
+  fill: var(--primary-filters-area-color);
+}
+
+.cell:nth-child(2) {
+  width: 7.2%;
+}
+
+.cell:nth-child(3) {
+  width: 11.7%;
+}
+
+.cell:nth-child(4) {
+  width: 11.7%;
+}
+
+.cell:nth-child(5) {
+  width: 6.4%;
+}
+
+.cell:nth-child(6) {
+  width: 8.5%;
+}
+
+.cell:nth-child(7) {
+  width: 52.4%;
+}
+
+.row:hover {
+  background-color: var(--secondary-color);
+}
+
+.row:last-child {
+  border-bottom: none;
+}
+
+.cell_canseled {
+  opacity: 0.5;
+}
+
+.text_completed {
+  color: var(--completed-icon-color);
+}
+
+::-webkit-scrollbar {
+  width: 0.25rem;
+}
+
+::-webkit-scrollbar-track {
+  margin: 0.25rem 0;
+  width: 0.25rem;
+  border-radius: 0.5rem;
+  background: rgba(0, 0, 0, 0.1);
+}
+
+::-webkit-scrollbar-thumb {
+  width: 0.25rem;
+  border-radius: 0.5rem;
+  background: rgba(0, 0, 0, 0.34);
+}
+
+.buttons {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.bunch {
+  color: var(--primary-color);
+}
+
+.overlayButton:hover {
+  background-color: var(--primary-color);
+}
+
+.overlayButton:hover span {
+  color: var(--primary-background-color);
+}
+
+.dropdown {
+  position: relative;
+}
+
+.dropdown_delete {
+  position: absolute;
+  left: 0;
+  top: -7.25rem;
+  display: flex;
+}
+
+.pages {
+  display: flex;
+  gap: 1.5rem;
+}
+
+.footer__pagination {
+  display: flex;
+  gap: 0.25rem;
+}
