+++ src/shared/ui/Table/TableBody/TableBody.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import classNames from 'classnames';
+import styles from './TableBody.module.css';
+
+export function TableBody({ className, children }) {
+  return <div className={classNames(styles._, className)}>{children}</div>;
+}
