+++ src/shared/ui/Table/TableBody/TableBody.module.css
@@ -1,0 +1,7 @@
+._ {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  background-color: var(--primary-background-color);
+  overflow: auto;
+}
