+++ src/shared/ui/Table/TableFooter/TableFooter.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import classNames from 'classnames';
+import styles from './TableFooter.module.css';
+
+export function TableFooter({ className, children }) {
+  return <div className={classNames(styles._, className)}>{children}</div>;
+}
