+++ src/shared/ui/Table/TableFooter/TableFooter.module.css
@@ -1,0 +1,6 @@
+._ {
+  display: flex;
+  justify-content: space-between;
+  padding: 0.5rem 1rem;
+  background-color: var(--primary-filters-area-color);
+}
