+++ src/shared/ui/Table/TableHeader/TableHeader.module.css
@@ -1,0 +1,3 @@
+._ {
+  background-color: var(--primary-color);
+}
