+++ src/shared/ui/Table/TableSortingCell/TableSortingCell.module.css
